Move NFT files to other folder
