+++ ui/components/NFTS_update/NFTListCurrentWallet.tsx
@@ -1,7 +1,7 @@
 import React, { ReactElement } from "react"
 import { selectCurrentAccountNFTs } from "@tallyho/tally-background/redux-slices/selectors"
 import NFTList from "./NFTList"
-import { useBackgroundSelector } from "../../../hooks"
+import { useBackgroundSelector } from "../../hooks"
 
 export default function NFTListCurrentWallet(): ReactElement {
   const collections = useBackgroundSelector(selectCurrentAccountNFTs)
