+++ ui/components/NFTS_update/NFTListPortfolio.tsx
@@ -1,7 +1,7 @@
 import React, { ReactElement } from "react"
 import { selectNFTCollections } from "@tallyho/tally-background/redux-slices/selectors"
 import NFTList from "./NFTList"
-import { useBackgroundSelector } from "../../../hooks"
+import { useBackgroundSelector } from "../../hooks"
 
 export default function NFTListPortfolio(): ReactElement {
   const collections = useBackgroundSelector(selectNFTCollections)
